updated tests

+++ src/main/resources/db/migration/V3__Create_programsparticipants_table.sql
@@ -1,6 +1,6 @@
 create table PROGRAMS_PARTICIPANTS (
     ID serial,
-    PROGRAM_ID int NOT NULL,
-    PARTICIPANT_ID int NOT NULL
+    PROGRAM_ID bigint NOT NULL,
+    PARTICIPANT_ID bigint NOT NULL
 );
